+++ cointop/portfolio.go
@@ -14,14 +14,14 @@ import (
 func (ct *Cointop) togglePortfolio() error {
 	ct.filterByFavorites = false
 	ct.portfoliovisible = !ct.portfoliovisible
-	ct.updateTable()
+	go ct.updateTable()
 	return nil
 }
 
 func (ct *Cointop) toggleShowPortfolio() error {
 	ct.filterByFavorites = false
 	ct.portfoliovisible = true
-	ct.updateTable()
+	go ct.updateTable()
 	return nil
 }