+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.4] - 2019-04-21
+### Changed
+- CoinMarketCap legacy V2 API to Pro V1 API
+
+### Added
+- Config option to use CoinMarketCap Pro V1 API KEY
+
+## [1.1.3] - 2019-02-25
+### Fixed
+- Vendor dependencies
+
 ## [1.1.2] - 2018-12-30
 ### Fixed
 - Paginate CoinMarketCap V1 API responses due to their backward-incompatible update